+++ c/frame/py_frame.h
@@ -1,9 +1,17 @@
 //------------------------------------------------------------------------------
-// This Source Code Form is subject to the terms of the Mozilla Public
-// License, v. 2.0. If a copy of the MPL was not distributed with this
-// file, You can obtain one at http://mozilla.org/MPL/2.0/.
+// Copyright 2018 H2O.ai
 //
-// © H2O.ai 2018
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
 //------------------------------------------------------------------------------
 #ifndef dt_FRAME_PYFRAME_h
 #define dt_FRAME_PYFRAME_h
@@ -58,6 +66,8 @@ class Frame : public PyObject {
     void m__dealloc__();
     void m__get_buffer__(Py_buffer* buf, int flags) const;
     void m__release_buffer__(Py_buffer* buf) const;
+    oobj m__getitem__(obj item);
+    void m__setitem__(obj item, obj value);
 
     oobj get_ncols() const;
     oobj get_nrows() const;
@@ -87,6 +97,11 @@ class Frame : public PyObject {
     void _replace_names_from_map(py::odict);
     Error _name_not_found_error(const std::string& name);
 
+    // getitem / setitem support
+    oobj _fast_getset(obj item, obj value);
+    oobj _main_getset(obj item, obj value);
+    oobj _fallback_getset(obj item, obj value);
+
     friend void pydatatable::_clear_types(pydatatable::obj*); // temp
     friend PyObject* pydatatable::check(pydatatable::obj*, PyObject*); // temp
     friend class FrameInitializationManager;
@@ -96,7 +111,7 @@ class Frame : public PyObject {
 
 extern PyObject* Frame_Type;
 extern PyObject* fread_fn;
-
+extern PyObject* fallback_makedatatable;
 
 }  // namespace py